# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fill.me API
  description: |
    Fill.me provides a comprehensive REST API for form creation, management, and analytics. 
    This API enables developers to integrate form functionality into their applications and manage form data programmatically.
    
    ## Authentication
    
    Fill.me API supports multiple authentication methods:
    - **JWT Token**: `Authorization: Bearer YOUR_JWT_TOKEN`
    - **API Key**: `X-API-Key: YOUR_API_KEY`
    - **Basic Auth**: `Authorization: Basic base64(email:password)`
    
    ## Rate Limiting
    
    Rate limits vary by subscription plan:
    - Free: 100 requests/hour
    - Starter: 500 requests/hour
    - Professional: 2,000 requests/hour
    - Enterprise: 10,000 requests/hour
    
  version: 1.3.0
  contact:
    name: Fill.me Support
    email: support@fill.me
    url: https://fill.me/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.fill.me
    description: Production server
  - url: http://localhost:3001/api
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
  - BasicAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BasicAuth:
      type: http
      scheme: basic
  schemas:
...
    Workspace:
      type: object
      properties:
        id:
          type: string
          example: "workspace_123"
        name:
          type: string
          example: "Marketing Team"
        description:
          type: string
        owner:
          type: string
        members:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
              role:
                type: string
                enum: [viewer, editor, admin]
              joinedAt:
                type: string
                format: date-time
        settings:
          type: object
          properties:
            isPublic:
              type: boolean
            allowInvitations:
              type: boolean
            defaultRole:
              type: string
        stats:
          type: object
          properties:
            totalForms:
              type: integer
            totalMembers:
              type: integer
            totalResponses:
              type: integer
...
paths:
...
  /workspaces:
    get:
      tags: [Workspaces]
      summary: List workspaces
      responses:
        '200':
          description: Workspaces retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'

    post:
      tags: [Workspaces]
      summary: Create workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                settings:
                  type: object
      responses:
        '201':
          description: Workspace created successfully
  /workspaces/{workspaceId}/invite:
    post:
      tags: [Workspaces]
      summary: Invite member to workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [viewer, editor, admin]
      responses:
        '201':
          description: Invitation sent successfully

  # Analytics
...